Refactor getIndentObjectRange into IndentObjectMatch
Also changes IndentObjectMatch to subclass TextObjectMovement
